requirements.txt
# -e git://github.com/Rapptz/discord.py@rewrite#egg=discord.py
discord.py
requests
python-dotenv